feat: moving from next to vite to diversify
