--- next-hardhat/.env.example
@@ -1,7 +1,0 @@
-DEPLOYER_PRIVATE_KEY="0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
-# hardhat acct #0 key, please update
-MUMBAI_DEPLOYER_PRIVATE_KEY="0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
-MUMBAI_ALCHEMY_KEY=""
-# hardhat acct #0 key, please update
-SEPOLIA_DEPLOYER_PRIVATE_KEY="0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
-SEPOLIA_ALCHEMY_KEY=""
