--- next-hardhat/.gitignore
@@ -1,55 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-.env
-.tmp
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/out/
-
-# production
-/build
-
-# misc
-.DS_Store
-*.pem
-
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
-
-# local env files
-.env*.local
-
-# vercel
-.vercel
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
-
-# hardhat
-cache
-
-# artifacts
-typechain-types
-proofs
-
-# noir
-crs
-
-# other
-.vscode
-.DS_Store
-artifacts
-.yarn/
