--- next-hardhat/.prettierignore
@@ -1,4 +1,0 @@
-# Add files here to ignore them from prettier formatting
-
-/dist
-/coverage
