--- next-hardhat/.prettierrc
@@ -1,7 +1,0 @@
-{
-    "arrowParens": "avoid",
-    "singleQuote": true,
-    "trailingComma": "all",
-    "printWidth": 100,
-    "proseWrap": "always"
-}
