--- next-hardhat/circuits/Nargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "noirstarter"
-type = "bin"
-authors = [""]
-compiler_version = "0.11.0"
-
-[dependencies]
