--- next-hardhat/circuits/Prover.toml
@@ -1,2 +1,0 @@
-x = "1"
-y = "2"
