--- next-hardhat/circuits/Verifier.toml
@@ -1,1 +1,0 @@
-y = "0x0000000000000000000000000000000000000000000000000000000000000002"
