--- next-hardhat/circuits/src/main.nr
@@ -1,3 +1,0 @@
-fn main(x : Field, y : pub Field) {
-    assert(x != y);
-}
