--- next-hardhat/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "noir-starter",
-  "version": "1.0.0",
-  "description": "A template repository to get started with writing zero knowledge programs with Noir.",
-  "scripts": {
-    "dev": "hardhat compile && yarn deploy && next",
-    "build": "hardhat compile && yarn deploy && next build",
-    "start": "next start",
-    "deploy": "hardhat compile && hardhat run --network ${NETWORK} scripts/deploy.ts",
-    "test": "hardhat test"
-  },
-  "dependencies": {
-    "@noir-lang/backend_barretenberg": "^0.17.0",
-    "@noir-lang/noir_js": "^0.17.0",
-    "@noir-lang/noir_wasm": "^0.17.0",
-    "@noir-lang/source-resolver": "^0.17.0",
-    "@noir-lang/types": "^0.17.0",
-    "@nomicfoundation/hardhat-chai-matchers": "^1.0.6",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.0",
-    "@nomicfoundation/hardhat-toolbox": "^2.0.1",
-    "@nomiclabs/hardhat-ethers": "^2.2.2",
-    "@nomiclabs/hardhat-etherscan": "^3.0.0",
-    "@typechain/ethers-v5": "^10.1.0",
-    "@typechain/hardhat": "^6.1.2",
-    "dotenv": "^16.0.3",
-    "hardhat": "^2.17.0",
-    "next": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-toastify": "^9.1.1",
-    "typechain": "^8.1.0"
-  },
-  "devDependencies": {
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.15.5",
-    "@types/react": "^18.0.26",
-    "chai": "^4.2.0",
-    "hardhat-gas-reporter": "^1.0.9",
-    "solidity-coverage": "^0.8.5",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.3"
-  }
-}
