--- next-hardhat/pages/App.css
@@ -1,15 +1,0 @@
-body {
-  margin: 0;
-}
-
-.faded {
-  opacity: 0.5;
-}
-
-.spinner {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  /* bring your own prefixes */
-  transform: translate(-50%, -50%);
-}
