--- next-hardhat/pages/index.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Component from '../components/component';
-
-export default function Page() {
-  return <Component />;
-}
