--- next-hardhat/utils/addresses.json
@@ -1,1 +1,0 @@
-{"chainId":31337,"verifier":"0x4A679253410272dd5232B3Ff7cF5dbB88f295319"}
