+++ vite-hardhat/.prettierrc
@@ -1,0 +1,7 @@
+{
+    "arrowParens": "avoid",
+    "singleQuote": true,
+    "trailingComma": "all",
+    "printWidth": 100,
+    "proseWrap": "always"
+}
