+++ vite-hardhat/circuits/Nargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "noirstarter"
+type = "bin"
+authors = [""]
+compiler_version = "0.19.4"
+
+[dependencies]
