+++ vite-hardhat/circuits/Prover.toml
@@ -1,0 +1,2 @@
+x = "1"
+y = "2"
