+++ vite-hardhat/circuits/Verifier.toml
@@ -1,0 +1,1 @@
+y = "0x0000000000000000000000000000000000000000000000000000000000000002"
