+++ vite-hardhat/circuits/src/main.nr
@@ -1,0 +1,3 @@
+fn main(x : Field, y : pub Field) {
+    assert(x != y);
+}
