+++ vite-hardhat/test/index.ts
@@ -1,0 +1,54 @@
+import { expect } from 'chai';
+import hre from 'hardhat';
+
+import { Noir } from '@noir-lang/noir_js';
+import { BarretenbergBackend } from '@noir-lang/backend_barretenberg';
+
+import { CompileResult, CompiledProgram, compile } from '@noir-lang/noir_wasm';
+import path from 'path';
+import { ProofData } from '@noir-lang/types';
+
+const getCircuit = async (name: string) => {
+  const compiled = await compile(path.resolve('circuits', 'src', `${name}.nr`));
+  return compiled;
+};
+
+describe('It compiles noir program code, receiving circuit bytes and abi object.', () => {
+  let noir: Noir;
+  let correctProof: ProofData;
+
+  before(async () => {
+    const compiled = await getCircuit('main');
+    const verifierContract = await hre.ethers.deployContract('UltraVerifier');
+
+    const verifierAddr = await verifierContract.deployed();
+    console.log(`Verifier deployed to ${verifierAddr.address}`);
+
+    const backend = new BarretenbergBackend(compiled.program);
+    noir = new Noir(compiled.program, backend);
+  });
+
+  it('Should generate valid proof for correct input', async () => {
+    const input = { x: 1, y: 2 };
+    // Generate proof
+    correctProof = await noir.generateFinalProof(input);
+    expect(correctProof.proof instanceof Uint8Array).to.be.true;
+  });
+
+  it('Should verify valid proof for correct input', async () => {
+    const verification = await noir.verifyFinalProof(correctProof);
+    expect(verification).to.be.true;
+  });
+
+  it('Should fail to generate valid proof for incorrect input', async () => {
+    try {
+      const input = { x: 1, y: 1 };
+      const incorrectProof = await noir.generateFinalProof(input);
+    } catch (err) {
+      // TODO(Ze): Not sure how detailed we want this test to be
+      expect(err instanceof Error).to.be.true;
+      const error = err as Error;
+      expect(error.message).to.contain('Cannot satisfy constraint');
+    }
+  });
+});
