+++ vite-hardhat/utils/addresses.json
@@ -1,0 +1,1 @@
+{"chainId":31337,"verifier":"0x8A791620dd6260079BF849Dc5567aDC3F2FdC318"}
